ZynAddSubFX/FLTK: fixed "long long" types
Replaced "long long" types by simple "long" or "int".

+++ plugins/zynaddsubfx/fltk/src/fl_open_uri.cxx
@@ -114,7 +114,7 @@ fl_open_uri(const char *uri, char *msg, int msglen) {
 #ifdef WIN32
   if (msg) snprintf(msg, msglen, "open %s", uri);
 
-  return (long long)ShellExecute(HWND_DESKTOP, "open", uri, NULL, NULL, SW_SHOW) > 32;
+  return (int)ShellExecute(HWND_DESKTOP, "open", uri, NULL, NULL, SW_SHOW) > 32;
 
 #elif defined(__APPLE__)
   char	*argv[3];			// Command-line arguments
